refactor: 주문테이블 루트 어그리거트로 패키지를 분리하라
