+++ src/main/java/kitchenpos/eatinorders/ordertable/application/OrderTableService.java
@@ -1,9 +1,9 @@
-package kitchenpos.eatinorders.application;
+package kitchenpos.eatinorders.ordertable.application;
 
 import kitchenpos.eatinorders.domain.OrderRepository;
 import kitchenpos.eatinorders.domain.OrderStatus;
-import kitchenpos.eatinorders.domain.OrderTable;
-import kitchenpos.eatinorders.domain.OrderTableRepository;
+import kitchenpos.eatinorders.ordertable.domain.OrderTable;
+import kitchenpos.eatinorders.ordertable.domain.OrderTableRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
