+++ src/test/java/kitchenpos/eatinorders/application/InMemoryOrderRepository.java
@@ -3,7 +3,7 @@ package kitchenpos.eatinorders.application;
 import kitchenpos.eatinorders.domain.Order;
 import kitchenpos.eatinorders.domain.OrderRepository;
 import kitchenpos.eatinorders.domain.OrderStatus;
-import kitchenpos.eatinorders.domain.OrderTable;
+import kitchenpos.eatinorders.ordertable.domain.OrderTable;
 
 import java.util.*;
